Fix to categoriesAdmin and ImagesAdmin

+++ src/components/ui/AdminComponents/CategoriesAdmin/CategoriesAdmin.tsx
@@ -9,12 +9,14 @@ import { useStoreCategory } from '../../../../store/useStoreCategory'
 import { useStoreModal } from '../../../../store/useStoreModal'
 import { ICategory } from '../../../../types/ICategory'
 import { AdminCategory } from '../../Modals/AdminCategory/AdminCategory'
+import useStoreProduct from '../../../../store/useStoreProduct'
 
 export const CategoriesAdmin = () => {
 
     
     const {categories, setCategories, fetchCategory, setActiveCategory} = useStoreCategory()
     const {openModalAdminCategory, modalAdminCategory} = useStoreModal()
+    const {products} = useStoreProduct()
 
     useEffect(() => {
         const getCategories = async() => {
@@ -29,6 +31,13 @@ export const CategoriesAdmin = () => {
     // Funcion para eliminar una categoria
     const handleDelete = async (idCategory : number) => {
         try {
+            const existInProduct = products.map(product =>
+                product.category.some(category => category.id === idCategory))
+
+            if (existInProduct){ 
+                errorAlert('Error', 'La categoria esta asociada a un producto')
+                return
+            }
             const deletedCategory = await deleteCategory(idCategory)
 
             succesAlert('Eliminado', 'La categoria se elimino correctamente')
